Rename ChannelType.ts to SpectralDefinition.ts
