+++ src/models/SpectralDefinition.ts
